# routes/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DEVICES API",
    "description": "This is docs for the devices and management api",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["AUTH"],
        "summary": "Login endpoint",
        "description": "Hit this and get access and refresh tokens for Login",
        "operationId": "loginUser",
        "requestBody": {
          "$ref": "#/components/requestBodies/Login"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Login"
          },
          "400": {
            "description": "Bad request on password mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "InvalidPassowordErr": {
                    "value": {
                      "message": "Invalid Password"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "NotFoundErr": {
                    "value": {
                      "message": "User Not Found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["USERS"],
        "summary": "Fetches all the users",
        "description": "Fetches all the users",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "To fetch users bases on their name",
            "example": "poco"
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched all users OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDetail"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["USERS"],
        "description": "This will create a new user in db with email, password and name",
        "summary": "Create a new user",
        "operationId": "createUser",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 3,
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 50,
                    "pattern": "^[a-zA-Z\\s]*$"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^[\\w-\\.]+@([\\w-]+\\.)+[\\w-]{2,4}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "isAdmin": {
                    "type": "boolean"
                  }
                },
                "required": ["name", "email", "password"],
                "example": {
                  "name": "Pugazh",
                  "password": "Password",
                  "email": "pugazh@gmail.com"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success creating a user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetail"
                },
                "examples": {
                  "User1": {
                    "$ref": "#/components/examples/CreateUserResponseExample"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/{id}": {
      "delete": {
        "description": "Provide params id to soft delete user",
        "summary": "Delete user from db",
        "operationId": "deleteUserById",
        "tags": ["USERS"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/QueryIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Success deleting User",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "example": "User Deleted successfully"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/change-password": {
      "post": {
        "description": "Change password only if the old password provided in body is valid",
        "summary": "Change user password",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": ["USERS"],
        "operationId": "updateUserPassword",
        "requestBody": {
          "required": true,
          "description": "Request body for user password updation",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oldPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                },
                "required": ["oldPassword", "newPassword"],
                "minProperties": 2,
                "maxProperties": 2,
                "additionalProperties": false
              },

              "examples": {
                "PasswordUpdateExample": {
                  "value": {
                    "oldPassword": "testpassword",
                    "newPassword": "Password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "400": {
            "description": "Bad request on password mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "InvalidPassowordErr": {
                    "value": {
                      "message": "Invalid Password"
                    }
                  }
                }
              }
            }
          },
          "200": {
            "description": "Successful password updation",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "Successful password Updation": {
                    "value": "Password Updated"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/devices": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": ["DEVICES"],
        "summary": "Fetches all devices",
        "description": "Must need the accesstoken then provide a array of devices in response",
        "operationId": "getDevices",
        "parameters": [
          {
            "name": "q",
            "description": "Provides the device name that is of any character for fuzzy search",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "Poco"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceDetail"
                  },
                  "example": [
                    {
                      "id": 1,
                      "name": "Poco",
                      "model": "X6",
                      "status": "GOOD",
                      "created_at": "2024-11-11T12:04.32.432123"
                    },
                    {
                      "id": 2,
                      "name": "Poco",
                      "model": "X6",
                      "status": "GOOD",
                      "created_at": "2024-11-11T12:04.32.432123"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Provide name, model and optional status GOOD OR DEFECT in body to create a device only admin can do this",
        "summary": "Create a device to db",
        "tags": ["DEVICES"],
        "operationId": "insertDevice",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "DeviceBodyToInsert": {
                  "value": {
                    "name": "Poco",
                    "model": "X6 Pro",
                    "status": "GOOD"
                  }
                }
              },
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "minProperties": 2,
                "maxProperties": 3,
                "required": ["name", "model"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 50
                  },
                  "model": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 50
                  },
                  "status": {
                    "type": "string",
                    "enum": ["GOOD", "DEFECT"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "201": {
            "description": "Device inserted in db successfully",
            "content": {
              "application/json": {
                "examples": {
                  "DeviceCreatedExRes": {
                    "$ref": "#/components/examples/DeviceDetailExample"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/DeviceDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/devices/instock": {
      "get": {
        "description": "This will provide array of devices that are not currently owned by the user and is not DEFECT",
        "summary": "Provides list of devices instock",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": ["DEVICES"],
        "operationId": "getInstockDevices",
        "responses": {
          "200": {
            "$ref": "#/components/responses/DevicesDetailsList"
          }
        }
      }
    },
    "/devices/assign": {
      "post": {
        "description": "This will check for device is present in db, then is it available to rent not assigned to other user and then will assign to the user based on accesstoken",
        "summary": "Assign a device",
        "operationId": "assignDevice",
        "tags": ["DEVICES"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "AdminAssignDeviceToUser": {
                  "description": "Only if admin is logged in and also the userId provided in body, else it will change the admin entry",
                  "value": {
                    "deviceId": 2,
                    "userId": 1,
                    "reason": "PERSONAL USE"
                  }
                },
                "UserAssigningDeviceToOwn": {
                  "description": "This will assign device to the user",
                  "value": {
                    "deviceId": 1,
                    "reason": "WFH"
                  }
                }
              },
              "schema": {
                "type": "object",
                "minProperties": 2,
                "maxProperties": 3,
                "properties": {
                  "deviceId": {
                    "type": "integer"
                  },
                  "userId": {
                    "description": "Provided if the admin wants to change any of the user else if admin it will change his own device status likewise",
                    "type": "integer"
                  },
                  "reason": {
                    "$ref": "#/components/schemas/ReasonEnum"
                  }
                },
                "required": ["deviceId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device rented successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "Device Rented"
              }
            }
          },
          "400": {
            "description": "If device is currently owned by others",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                },
                "example": {
                  "message": "Device is taken"
                }
              }
            }
          },
          "404": {
            "description": "Invalid Device Id will make device not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                },
                "example": {
                  "message": "Device not found"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
      "QueryIdParam": {
        "name": "id",
        "in": "path",
        "description": "Provide **Particular ID**\n",
        "required": true,
        "schema": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "requestBodies": {
      "Login": {
        "required": true,
        "description": "Auth Login request body contents",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LoginBody"
            },
            "examples": {
              "LoginRequestBody": {
                "$ref": "#/components/examples/LoginBody"
              }
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorSchema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "ReasonEnum": {
        "type": "string",
        "enum": ["PERSONAL USE", "AT SERVICE CENTER", "WFH", "RETURNED", "TESTING"],
        "default": "PERSONAL USE"
      },
      "LoginBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^[\\w-\\.]+@([\\w-]+\\.)+[\\w-]{2,4}$"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": ["email", "password"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": ["message", "accessToken", "refreshToken"]
      },
      "DeviceDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "status": {
            "enum": ["GOOD", "DEFECT"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "responses": {
      "Login": {
        "description": "Login Response 200 success",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LoginResponse"
            },
            "examples": {
              "LoginResponse": {
                "$ref": "#/components/examples/LoginResponse"
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Throwed when no token is found i think so **Unauthorized**\n",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "default": "Unauthorized"
                }
              }
            }
          }
        }
      },
      "DevicesDetailsList": {
        "description": "Device list based on the endpoint",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/DeviceDetail"
              }
            },
            "examples": {
              "DevicesExample": {
                "$ref": "#/components/examples/DeviceDetailsExample"
              }
            }
          }
        }
      }
    },
    "examples": {
      "LoginBody": {
        "value": {
          "email": "aakashit13@gmail.com",
          "password": "password"
        }
      },
      "LoginResponse": {
        "value": {
          "message": "Login success",
          "accessToken": "RandomUUID",
          "refreshToken": "RandomUUID"
        }
      },
      "CreateUserResponseExample": {
        "value": {
          "id": 1,
          "name": "Aakash",
          "password": "RANDOM UUID",
          "email": "aakash@gmail.com",
          "created_at": "2024-11-11T12:04.32.432123"
        }
      },
      "DeviceDetailExample": {
        "value": {
          "id": 2,
          "name": "Poco",
          "model": "X6",
          "status": "GOOD",
          "created_at": "2024-11-11T12:04.32.432123"
        }
      },
      "DeviceDetailsExample": {
        "value": [
          {
            "id": 1,
            "name": "Poco",
            "model": "X6",
            "status": "GOOD",
            "created_at": "2024-11-11T12:04.32.432123"
          },
          {
            "id": 2,
            "name": "Poco",
            "model": "X6",
            "status": "GOOD",
            "created_at": "2024-11-11T12:04.32.432123"
          }
        ]
      },
...
    }
  }
}